Refactoring, phase I
* restructure the project
* move business logic into separate module
* reorganize code into functions
