+++ srv/webapp/video_streaming/utils/normalize_image.py
